requirements.txt
Flask>=2.3.2,<2.4.0
Flask-Login>=0.6.2,<0.7.0
Flask-Migrate>=4.0.4,<4.1.0
Flask-SQLAlchemy>=3.0.5,<3.1.0
Flask-WTF>=1.1.1,<1.2.0
psycopg2-binary>=2.9.6
Werkzeug>=2.3.6,<2.4.0
email-validator>=2.0.0,<2.1.0
python-dotenv>=1.0.0,<1.1.0
Flask-Bcrypt>=1.0.1,<1.1.0
gunicorn>=20.1.0,<20.2.0 